requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
omegaconf>=2.3.0
einops>=0.6.1
transformers>=4.30.0
pycocotools>=2.0.6
pytorch-lightning>=2.0.0
torchmetrics>=0.11.4
kornia>=0.6.12
matplotlib>=3.7.0
pyyaml>=6.0
scipy>=1.10.0
